+++ packages/tinacms/src/tina-cms.tsx
@@ -232,7 +232,9 @@ export const TinaCMSProvider2 = ({
         isLocalClient={isLocalClient}
         cmsCallback={props.cmsCallback}
         mediaStore={props.mediaStore}
-        schema={schema}
+        apiUrl={apiURL}
+        // Not ideal but we need this for backwards compatibility for now. We can clean this up when we require a config.{js,ts} file
+        schema={{ ...schema, config: { ...schema.config, ...props } }}
       >
         <style>{styles}</style>
         <ErrorBoundary>